Mappers, config files, renames, new enums
