+++ src/main/resources/mappers/LicencseMapper.xml
@@ -9,9 +9,20 @@
         <result columnn = "EXPIRATION_DATE" property "expirationDate"/> 
     </resultMap>
 
-
     <select id ="getById" resultMap = "LicenseResultMap">
         SELECT * FROM Licenses WHERE ID = #{id}
+    </select>
+
+    <select id="getByNumber" resultMap="LicenseResultMap">
+        SELECT * FROM Licenses WHERE NUMBER = #{number}
+    </select>
+
+    <select id="getExpiredLicenses" resultMap="LicenseResultMap">
+        SELECT * FROM Licenses WHERE EXPIRATION_DATE < CURDATE();
+    </select>
+
+    <select id="getValidLicenses" resultMap="LicenseResultMap">
+        SELECT * FROM Licenses WHERE EXPIRATION_DATE >= CURDATE();
     </select>
 
     <insert id="save" parameterType="com.solvd.socialnetwork.models.License" useGeneratedKeys="true" keyProperty="id">
