+++ src/main/resources/mybatis-config.xml
@@ -22,9 +22,8 @@
         <mapper resource ="mappers/DriverMapper.xml"/>
         <mapper resource ="mappers/LicenseMapper.xml"/>
         <mapper resource ="mappers/LocationMapper.xml"/>
-        <mapper resource ="mappers/PersonMapper.xml"/>
         <mapper resource ="mappers/RouteMapper.xml"/>
-        <mapper resource ="mappers/UserMapper.xml"/>
+        <mapper resource ="mappers/ClientMapper.xml"/>
         <mapper resource ="mappers/VehicleMapper.xml"/>
     </mappers>
